+++ packages/nimble-components/src/breadcrumb/tests/breadcrumb-matrix.stories.ts
@@ -12,7 +12,7 @@ const metadata: Meta = {
     parameters: {
         design: {
             artboardUrl:
-                'https://xd.adobe.com/view/8ce280ab-1559-4961-945c-182955c7780b-d9b1/screen/7b53bb3e-439b-4f13-9d5f-55adc7da8a2e/specs/'
+                'https://xd.adobe.com/view/33ffad4a-eb2c-4241-b8c5-ebfff1faf6f6-66ac/screen/7b53bb3e-439b-4f13-9d5f-55adc7da8a2e/specs/'
         },
         controls: { hideNoControlsWarning: true },
         a11y: { disabled: true }
@@ -21,22 +21,24 @@ const metadata: Meta = {
 
 export default metadata;
 
-type LinkHrefState = [string, string | null];
-const linkHrefStates: LinkHrefState[] = [
-    ['Link', parent.location.href],
-    ['Current (No Link)', null]
+type BreadcrumbStyleState = [string, string];
+const breadcrumbStyleStates: BreadcrumbStyleState[] = [
+    ['', ''],
+    [' (Prominent Links style)', 'prominent-links']
 ];
-
-const component = ([linkHrefName, href]: LinkHrefState): ViewTemplate => html`
-    <nimble-breadcrumb>
-        <nimble-breadcrumb-item href="${() => href}">
-            ${() => `Breadcrumb - ${linkHrefName}`}
+const component = ([
+    styleStateName,
+    style
+]: BreadcrumbStyleState): ViewTemplate => html`
+    <nimble-breadcrumb class="${() => style}" style="margin-right: 24px">
+        <nimble-breadcrumb-item href="${parent.location.href}">
+            ${() => `Breadcrumb${styleStateName} - Link`}
         </nimble-breadcrumb-item>
+        <nimble-breadcrumb-item>Current (No Link)</nimble-breadcrumb-item>
     </nimble-breadcrumb>
 `;
-
 export const breadcrumbThemeMatrix: Story = createRenderer(
-    themeWrapper(createMatrix(component, [linkHrefStates]))
+    themeWrapper(createMatrix(component, [breadcrumbStyleStates]))
 );
 
 export const hiddenBreadcrumb: Story = createRenderer(